requirements.txt
streamlit>=1.32.0
pandas>=2.0.3
anthropic>=0.18.0
python-dotenv>=1.0.0
tikapi>=1.0.0
Pillow>=10.4.0
openpyxl>=3.1.2
requests>=2.31.0
supabase>=2.0.0
websockets>=15.0.0